Address @benvinegar's suggestions.
